2nd commit
We came into a realization that the project does not need the full on validation system
that was introduced in the 1st commit so it was shrunk into two basic functions that only
check the length of each attribute.

Here we started creating an UI and finished most of the functions that interact with
the linked addresses' list.